DEV-99 Restructured FaultAndWarningLogger folders
