+++ src/components/FaultsAndWarningLogger/FaultsAndWarningLogger.module.scss
