+++ src/components/FaultsAndWarningLogger/FaultsAndWarningLogger.tsx
@@ -1,15 +1,13 @@
-import ConsoleList from "@components/MessageLogger/ConsoleList/ConsoleList";
+import MessageLogger from "@components/FaultsAndWarningLogger/MessageLogger/MessageLogger";
 import { useState } from "react";
-import { RootState } from "store";
-import { useSelector } from "react-redux";
-import { store } from "../../../store";
-import styles from "@components/MessageLogger/FaultsAndWarningList/FaultsAndWarningList.module.scss";
+import { store } from "../../store";
+import styles from "@components/FaultsAndWarningLogger/FaultsAndWarningLogger.module.scss";
 import { useInterval } from "@hooks/useInterval";
 import { Message } from "@models/Message";
 const warningColor = { h: 41, s: 100, l: 40, a: 1 };
 const faultColor = { h: 0, s: 100, l: 40, a: 1 };
 
-export const FaultsAndWarningList = () => {
+export const FaultsAndWarningLogger = () => {
     const [messages, setMessages] = useState({
         fault: [] as Message[],
         warning: [] as Message[],
@@ -24,12 +22,12 @@ export const FaultsAndWarningList = () => {
 
     return (
         <div className={styles.containerMessages}>
-            <ConsoleList
+            <MessageLogger
                 title={"Warnings"}
                 messages={messages.warning}
                 color={warningColor}
             />
-            <ConsoleList
+            <MessageLogger
                 title={"Faults"}
                 messages={messages.fault}
                 color={faultColor}
