+++ src/components/FaultsAndWarningLogger/MessageLogger/MessageList/MessageList.module.scss
@@ -7,6 +7,7 @@
     overflow-y: auto;
     background-color: styles.$isle-color;
     @include styles.code-text;
+    border-radius: styles.$normal-border-radius;
 
     > *:not(:last-child) {
         border-bottom: 1px solid black;
