+++ src/components/FaultsAndWarningLogger/MessageLogger/MessageLogger.module.scss
@@ -1,6 +1,6 @@
 @use "@styles/styles.scss";
 
-.containerMessages {
+.messageLogger {
     display: flex;
     flex-direction: column;
     width: 100%;
